+++ Dockerfile
@@ -17,6 +17,7 @@ ENV VPR_CHURCH_MEMBERS_DATABASE_URL $VPR_CHURCH_MEMBERS_DATABASE_URL \
     VPR_AUTH_TOKEN $VPR_AUTH_TOKEN \
     VPR_AUTH_AUD $VPR_AUTH_AUD \
     SCOPE $SCOPE \
+    SERVER $SERVER \
     APP_LANG $APP_LANG
 
 COPY --from=build /go/src/app/bundles/* ./bundles/
