+++ build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+GOOS=linux CGO_ENABLED=0 go build
+zip function.zip church-members-api
